Fix the issue with reopening the modal
While navigating the cart using the keyboard, there was an issue when a
user opened the modal and then attempted to close it. The keyboard hit
would close the modal and then instantly reopen it due to the focus on
the buy button. To address this, the focus is now manually shifted to
another element when opening the modal.